api/requirements.txt
click==8.1.7
fastapi==0.110.2
httpx==0.27.2
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
openai==1.23.6
pyzmq==26.0.2
uvicorn==0.29.0
PyPDF2==3.0.1
aiofiles==23.2.1
elasticsearch[async]
colorlog==6.8.2